requirements.txt
pandas
sqlalchemy
streamlit
pyyaml
